Seprate one Links resource to multiple resources for make things clean and Swagger work correctly
