+++ api.py
@@ -6,7 +6,10 @@ import json
 
 from resources.user.Login import Login
 from resources.user.Signup import Signup
-from resources.links.Links import Links
+from resources.links.LinksDeleter import LinksDeleter
+from resources.links.LinksAdder import LinksAdder
+from resources.links.LinksGetter import LinksGetter
+
 
 from db import db
 
@@ -49,8 +52,18 @@ api.add_resource(
     Signup, CONFIG.get('routes', {}).get('user', {}).get('signup')
 )
 api.add_resource(
-    Links, CONFIG.get('routes', {}).get('links', {}).get('main'),
+    LinksGetter,
+    CONFIG.get('routes', {}).get('links', {}).get('main'),
+    CONFIG.get('routes', {}).get('links', {}).get('get_id'),
+    endpoint='links'
+)
+api.add_resource(
+    LinksDeleter,
     CONFIG.get('routes', {}).get('links', {}).get('get_id')
+)
+api.add_resource(
+    LinksAdder,
+    CONFIG.get('routes', {}).get('links', {}).get('main')
 )
